refactor get_config_from_args to dagster._cli.workspace.cli_target
